Feature: staking (#146)
* Feature: staking

* Fix: parts of utia in rewards and commissions

* Fixes

* Fix: compute total stake in genesis

* Fix: unbonding via exec message

* Fix: nil pointer

* Fix: validator editing and burning on jail

* Fix: increase series limit

* Update celestia-app

* Fix: save staking log in genesis

* Fix: staking series

* Fix: api docs

* Fix: change delegations amount on redelegate

* Fix: change delegations amount on undelegate

* Fix: change delegations amount on cancel undelegate

* Fix: jail burned field name

* Fix: slashing constants

* Fix: update slashed delegations

* Add jailed filter

* Fix: security issus

+++ cmd/api/handler/requests.go
@@ -18,8 +18,8 @@ const (
 )
 
 type limitOffsetPagination struct {
-	Limit  uint64 `json:"limit"  param:"limit"  query:"limit"  validate:"omitempty,min=1,max=100"`
-	Offset uint64 `json:"offset" param:"offset" query:"offset" validate:"omitempty,min=0"`
+	Limit  int    `json:"limit"  param:"limit"  query:"limit"  validate:"omitempty,min=1,max=100"`
+	Offset int    `json:"offset" param:"offset" query:"offset" validate:"omitempty,min=0"`
 	Sort   string `json:"sort"   param:"sort"   query:"sort"   validate:"omitempty,oneof=asc desc"`
 }
 
@@ -33,8 +33,8 @@ func (p *limitOffsetPagination) SetDefault() {
 }
 
 type addressListRequest struct {
-	Limit  uint64 `query:"limit"  validate:"omitempty,min=1,max=100"`
-	Offset uint64 `query:"offset" validate:"omitempty,min=0"`
+	Limit  int    `query:"limit"  validate:"omitempty,min=1,max=100"`
+	Offset int    `query:"offset" validate:"omitempty,min=0"`
 	Sort   string `query:"sort"   validate:"omitempty,oneof=asc desc"`
 }
 
@@ -59,8 +59,8 @@ func pgSort(sort string) storage.SortOrder {
 }
 
 type txListRequest struct {
-	Limit           uint64      `query:"limit"             validate:"omitempty,min=1,max=100"`
-	Offset          uint64      `query:"offset"            validate:"omitempty,min=0"`
+	Limit           int         `query:"limit"             validate:"omitempty,min=1,max=100"`
+	Offset          int         `query:"offset"            validate:"omitempty,min=0"`
 	Sort            string      `query:"sort"              validate:"omitempty,oneof=asc desc"`
 	Height          uint64      `query:"height"            validate:"omitempty,min=1"`
 	Status          StringArray `query:"status"            validate:"omitempty,dive,status"`
